Moved transformCandleSticksForCharts to helper folder. Removed its use in reducer and use it in actionCreator insead.
