+++ src/helpers/transform-candlesticks-for-chart.js
